requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.9
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
elasticsearch==7.8.0
fake-useragent==0.1.11
geographiclib==1.50
geopy==2.0.0
googletransx==2.4.2
idna==2.10
multidict==4.7.6
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
Pillow==7.1.2
pycares==3.1.1
pycparser==2.20
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
tweepy==3.8.0
-e git+https://github.com/twintproject/twint.git@44273dd9c69cd5d5f2754b8e1d725f4ffa62311a#egg=twint
urllib3==1.25.9
yarl==1.4.2
